Final Project/Streamlit/requirements.txt
streamlit
streamlit-option-menu
joblib
numpy
huggingface-hub
torch
torchvision
Pillow
scikit-learn
pandas